Refactor: utils 폴더 js 하위폴더에서 src 하위폴더로 변경
